+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Getter for `::$components`
+
 ## [1.2.3] - 2018-03-21
 
 ### Changed